Replace unsigned int and int by NSUInteger and NSInteger
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@15134 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKAddCommand.m
@@ -78,7 +78,7 @@
         id locationSpecifier = [[self arguments] objectForKey:@"ToLocation"];
         id insertionContainer = nil;
         NSString *insertionKey = nil;
-        int insertionIndex = -1;
+        NSInteger insertionIndex = -1;
         NSScriptClassDescription *containerClassDescription = nil;
         NSArray *classDescriptions = [insertionObjects valueForKey:@"scriptClassDescription"];
         NSScriptClassDescription *insertionClassDescription = [classDescriptions containsObject:[NSNull null]] ? nil : [NSScriptClassDescription commonAncestorForClassDescriptions:classDescriptions];
