+++ BDSKAutofileCommand.m
@@ -55,8 +55,8 @@
     NSString *location = [params objectForKey:@"to"];
     NSNumber *checkNumber = [params objectForKey:@"check"];
 	BOOL check = checkNumber ? [checkNumber boolValue] : YES;
-    int mask = 0;
-    unsigned int i = indexNumber ? [indexNumber unsignedIntValue] - 1 : 0;
+    NSInteger mask = 0;
+    NSUInteger i = indexNumber ? [indexNumber unsignedIntValue] - 1 : 0;
     
 	if (pub == nil) {
 		[self setScriptErrorNumber:NSRequiredArgumentsMissingScriptError];
