+++ BDSKComplexStringEditor.h
@@ -46,12 +46,12 @@
     IBOutlet BDSKBackgroundView *backgroundView;
 	NSTableView *tableView;
     BDSKComplexStringFormatter *formatter;
-	int row;
-	int column;
+	NSInteger row;
+	NSInteger column;
     BOOL editable;
 }
 
-- (BOOL)attachToTableView:(NSTableView *)aTableView atRow:(int)aRow column:(int)aColumn withValue:(NSString *)aString formatter:(BDSKComplexStringFormatter *)aFormatter;
+- (BOOL)attachToTableView:(NSTableView *)aTableView atRow:(NSInteger)aRow column:(NSInteger)aColumn withValue:(NSString *)aString formatter:(BDSKComplexStringFormatter *)aFormatter;
 - (BOOL)isEditing;
 - (BOOL)isEditable;
 - (void)setEditable:(BOOL)flag;
